refactor: move root cookie sync to its own file
